GUACAMOLE-1479: Move disabled methods up to new Disableable interface.
